requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==2.0.12
Django==3.2
django-filter==23.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.26.0
six==1.16.0
SQLAlchemy==2.0.36
sqlparse==0.5.1
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
